requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
exceptiongroup==1.1.0
fire==0.4.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.4.0
pytest==7.2.0
python-dotenv==0.20.0
python-environ==0.4.54
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
termcolor==1.1.0
tomli==2.0.1
urllib3==1.26.12
whitenoise==6.2.0
